Adjust the stack upgrade confirmation message for add-on only apps
Apps that have never had a deploy (eg empty apps or those that are used
to hold add-ons only) switch stacks instantly without the need for a
re-deploy of the app (as of heroku/api#10754).

However when upgrading an app's stack via the CLI, the console output
previously always indicated a re-deploy was necessary, even for apps
where this is now not true.

This has been seen to cause confusion in support tickets.

As such, the redploy message is now only shown when the response from API
indicates that the upgrade is still pending (ie: when `build_stack`,
the requested stack, differs from the `stack`, the current app stack).

The `stacks:set` command has also been updated to use `cli.action()`,
to prevent the existing lack of output until the API request returns.

Refs W-7249171.